Fast-Neural-Style-Transfer/requirements.txt
torch==1.1.0
Flask==0.12.2
torchvision==0.4.0
numpy==1.17.4
tqdm==4.40.0
Pillow==7.1.2
